% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sunspot}
\alias{sunspot}
\title{Sunspot Dataset}
\format{
A data frame with ... observations and 16 variables.
}
\usage{
data(sunspot)
}
\description{
This dataset provides sunspot data published on NASA's Solar Physics website and
compiled by  Dr. David Hathaway, a retired member of the MSFC solar physics group
and NASA's Ames Research Center in California.
}
\details{
The dataset is created from ASCII text files containing records for individual
years of sunspot observations. Each file consists of records with information
on individual sunspot groups for each day that spots were observed.

The Royal Greenwich Observatory (RGO) compiled sunspot observations from a small
network of observatories to produce a dataset of daily observations starting in
May of 1874. The observatory concluded this dataset in 1976 after the US Air
Force (USAF) started compiling data from its own Solar Optical Observing Network
(SOON). This work was continued with the help of the US National Oceanic and
Atmospheric Administration (NOAA) with much of the same information being
compiled through to the present.

The dataset includes the following variables:
\describe{
\item{\code{year}}{Character, ...}
\item{\code{month}}{Character, ...}
\item{\code{day}}{Character, ...}
\item{\code{time}}{Character, ...}
\item{\code{group}}{Character, ...}
\item{\code{group_suffix}}{Character, ...}
\item{\code{group_type}}{Character, ...}
\item{\code{observed_umbral_area}}{Character, ...}
\item{\code{observed_whole_spot_area}}{Character, ...}
\item{\code{corrected_umbral_area}}{Character, ...}
\item{\code{corrected_whole_spot_area}}{Character, ...}
\item{\code{distance_from_center}}{Character, ...}
\item{\code{position_angle}}{Character, ...}
\item{\code{longitude}}{Character, ...}
\item{\code{latitude}}{Character, ...}
\item{\code{central_meridian_distance}}{Character, ...}
}
}
\references{
\url{https://solarscience.msfc.nasa.gov/greenwch.shtml}
}
\author{
Donnie Minnick \email{donnie.minnick@gmail.com}
}
\keyword{dataset}
